+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 ## [Unreleased]
 
+### Added
+
+- Input for specifying the type of the project’s version file.
+- Input for specifying path to the project’s version file.
+
 ### Changed
 
 - Place of the version upload to POST action.